Now geting mogoose url from either env variable or config, some string updates for single visits, added some logging for debugging
And added my site's styling so make it look a bit nicer to look at.

There's a bug where if I attempt to save session data using
mongo-session... there's either a delay in the session being saved, or
the save doesn't happen at all.... so the user is redirected back to the
index page in a non-logged in state. This is being tracked with
https://github.com/thisisthefuture/whatsfordinner/issues/3